JE2-68 updated password streght validation, registration logic moved to a hook
